TIP-860: Move AssociationType related api tests
