+++ tests/back/Pim/Structure/EndToEnd/AssociationType/ExternalApi/PartialUpdateListAssociationTypeEndToEnd.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\AssociationType;
+namespace AkeneoTest\Pim\Structure\EndToEnd\AssociationType\ExternalApi;
 
-use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\Stream\StreamResourceResponse;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class PartialUpdateListAssociationTypeIntegration extends ApiTestCase
+class PartialUpdateListAssociationTypeEndToEnd extends ApiTestCase
 {
     public function testCreateAndUpdateAListOfAssociationTypes()
     {
